requirements.txt
colorama
groq
python-dotenv
asyncio